[94] Colon, not comma
Not to be a stickler about this, but I do prefer a colon here, since
it's followed by a list of things that happened in Swift development,
and I think its usage here is grammatically correct. Let me know if
the first sentence needs to be rewritten altogether, but I think if
it's going to stay as-is then the colon isn't a bad choice here.